fix: map size, offset and geometry
The map texture is offset and aligned with the prime meridian to avoid
having to manually do so in code.

The image size has been reduced by 60% by not sampling the alpha, but rgb
channels instead.

The vertices for the spheres have been reduced from 4900 to 2880 by using an
icosahedron instead.